functions/trigger/functions/requirements.txt
firebase_functions~=0.1.0
# vertexai
google-cloud-aiplatform